[MTDSA-11900] Retrieve SA Tax Calc Alignments to Code and RAML (#386)
* MTDSA-11900 Retrieve SA Tax Calc Alignments to Code and RAML

* addressed comments

+++ app/v3/models/response/retrieveCalculation/calculation/reliefs/ForeignTaxCreditReliefDetail.scala
@@ -25,7 +25,7 @@ case class ForeignTaxCreditReliefDetail(incomeSourceType: Option[IncomeSourceTyp
                                         countryCode: String,
                                         foreignIncome: BigDecimal,
                                         foreignTax: Option[BigDecimal],
-                                        dtaRate: Option[BigInt],
+                                        dtaRate: Option[BigDecimal],
                                         dtaAmount: Option[BigDecimal],
                                         ukLiabilityOnIncome: Option[BigDecimal],
                                         foreignTaxCredit: BigDecimal,
